Move music match logic out of handler.py and into own file
